requirements.txt
flask
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
prometheus-client
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
Flask-SQLAlchemy
psycopg2-binary
bcrypt
pyjwt
opentelemetry-instrumentation-sqlalchemy